requirements.txt
pytest-mozwebqa==1.0
requests==0.13.3
